refactor: update AWS client implementations to use inner structs for better encapsulation
