+++ orchestrator/crates/orchestrator/src/core/client/storage/s3.rs
@@ -7,9 +7,29 @@ use aws_sdk_s3::Client;
 use bytes::Bytes;
 use std::sync::Arc;
 
+
+/// AWSS3 is a struct that represents an AWS S3 client.
+#[derive(Clone, Debug)]
+pub(crate) struct InnerAWSS3(pub(crate) Arc<Client>);
+
+impl InnerAWSS3 {
+    /// Creates a new instance of InnerAWSS3 with the provided AWS configuration.
+    /// # Arguments
+    /// * `aws_config` - The AWS configuration.
+    ///
+    /// # Returns
+    /// * `Self` - The new instance of InnerAWSS3.
+    pub fn new(aws_config: &SdkConfig) -> Self {
+        let mut s3_config_builder = aws_sdk_s3::config::Builder::from(aws_config);
+        s3_config_builder.set_force_path_style(Some(true));
+        let client = Client::from_conf(s3_config_builder.build());
+        Self(Arc::new(client))
+    }
+}
+
 #[derive(Clone, Debug)]
 pub struct AWSS3 {
-    pub(crate) client: Arc<Client>,
+    pub(crate) inner: InnerAWSS3,
     bucket_name: Option<String>,
 }
 
@@ -22,14 +42,15 @@ impl AWSS3 {
     /// # Returns
     /// * `Self` - The new instance of AWSS3.
     pub fn new(aws_config: &SdkConfig, args: Option<&StorageArgs>) -> Self {
-        let mut s3_config_builder = aws_sdk_s3::config::Builder::from(aws_config);
-        s3_config_builder.set_force_path_style(Some(true));
-        let client = Client::from_conf(s3_config_builder.build());
-        Self { client: Arc::new(client), bucket_name: args.map(|a| a.bucket_name.clone()) }
+        Self { inner: InnerAWSS3::new(aws_config), bucket_name: args.map(|a| a.bucket_name.clone()) }
     }
 
     pub(crate) fn bucket_name(&self) -> Result<String, StorageError> {
         self.bucket_name.clone().ok_or_else(|| StorageError::InvalidBucketName("Bucket name is not set".to_string()))
+    }
+    
+    pub(crate) fn client(&self) -> &Client {
+        self.inner.0.as_ref()
     }
 }
 
@@ -44,7 +65,7 @@ impl StorageClient for AWSS3 {
     /// * `Result<Bytes, StorageError>` - The result of the get operation.
     async fn get_data(&self, key: &str) -> Result<Bytes, StorageError> {
         // Note: unwrap is safe here because the bucket name is set in the constructor
-        let output = self.client.get_object().bucket(self.bucket_name()?).key(key).send().await?;
+        let output = self.client().get_object().bucket(self.bucket_name()?).key(key).send().await?;
 
         let data = output.body.collect().await.map_err(|e| StorageError::ObjectStreamError(e.to_string()))?;
         Ok(data.into_bytes())
@@ -59,7 +80,7 @@ impl StorageClient for AWSS3 {
     /// * `Result<(), StorageError>` - The result of the put operation.
     async fn put_data(&self, data: Bytes, key: &str) -> Result<(), StorageError> {
         // Note: unwrap is safe here because the bucket name is set in the constructor
-        self.client.put_object().bucket(self.bucket_name()?).key(key).body(data.into()).send().await?;
+        self.client().put_object().bucket(self.bucket_name()?).key(key).body(data.into()).send().await?;
         Ok(())
     }
 
@@ -71,6 +92,6 @@ impl StorageClient for AWSS3 {
     /// * `Result<(), StorageError>` - The result of the delete operation.
     async fn delete_data(&self, key: &str) -> Result<(), StorageError> {
         // Note: unwrap is safe here because the bucket name is set in the constructor
-        Ok(self.client.delete_object().bucket(self.bucket_name()?).key(key).send().await.map(|_| ())?)
+        Ok(self.client().delete_object().bucket(self.bucket_name()?).key(key).send().await.map(|_| ())?)
     }
 }
